# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  response.Response:
    properties:
      data: {}
      error: {}
    type: object
host: localhost:8080
info:
  contact:
    email: anhquoc18092003@gmail.com
    name: Tran Phuoc Anh Quoc
  description: Swagger API for Go Clean Architecture.
  license:
    name: MIT
    url: https://github.com/MartinHeinz/go-project-blueprint/blob/master/LICENSE
  termsOfService: http://swagger.io/terms/
  title: Ecommerce Clean Architecture Swagger API
  version: "1.0"
paths:
...
  /products:
    get:
      description: Fetches a paginated list of products based on the provided filter
        parameters.
      parameters:
      - description: Search keyword for products
        in: query
        name: search
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Number of items per page (default: 10)'
        in: query
        name: size
        type: integer
      - description: Field to sort by
        in: query
        name: order_by
        type: string
      - description: Sort in descending order (true/false)
        in: query
        name: order_desc
        type: boolean
      - description: Retrieve all products without pagination
        in: query
        name: take_all
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved the list of products
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request - Invalid query parameters
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error - An error occurred while processing
            the request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Retrieve a list of products
      tags:
      - Products
    post:
      consumes:
      - multipart/form-data
      description: Creates a new product based on the provided details.
      parameters:
      - description: Product Name
        in: formData
        name: name
        required: true
        type: string
      - description: Product Description
        in: formData
        name: description
        required: true
        type: string
      - description: Product Image
        in: formData
        name: image
        required: true
        type: file
      - description: Product Price (must be greater than 0)
        in: formData
        name: price
        required: true
        type: number
      produces:
      - application/json
      responses:
        "201":
          description: Product created successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request - Invalid parameters
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized - User not authenticated
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden - User does not have the required permissions
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict - Code or Name already in use
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error - An error occurred while processing
            the request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Create a new product
      tags:
      - Products
  /products/{id}:
    delete:
      description: Deletes an existing product by its ID.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Product deleted successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request - Invalid parameters
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized - User not authenticated
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden - User does not have the required permissions
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found - Product with the specified ID not found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error - An error occurred while processing
            the request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete a product
      tags:
      - Products
    get:
      description: Fetches the details of a specific product based on the provided
        product ID.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved the product
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request - Invalid product ID
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized - User not authenticated
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden - User does not have the required permissions
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found - Product with the specified ID not found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error - An error occurred while processing
            the request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Retrieve a product by its ID
      tags:
      - Products
    put:
      consumes:
      - multipart/form-data
      description: Update an existing product based on the provided details.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Product Name
        in: formData
        name: name
        type: string
      - description: Product Description
        in: formData
        name: description
        type: string
      - description: Product Image
        in: formData
        name: image
        type: file
      - description: Product Price (must be greater than or equal to 0)
        in: formData
        name: price
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: Product updated successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request - Invalid parameters
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized - User not authenticated
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden - User does not have the required permissions
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found - Product with the specified ID not found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error - An error occurred while processing
            the request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Update a product
      tags:
      - Products
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"